refactor(130): split useDrumGrid 613-line hook into 3 focused hooks
useDrumGrid.ts: 613 → 238 lines (-61%) — now a thin coordinator
New src/hooks/useGridDragPaint.ts: 211 lines
  - Drag/paint state, touch state, applyDragAction
  - handleMouseDown, handleMouseEnter, handleTouchStart/Move/End
  - Global mouseup effect (ends drag on outside release)
New src/hooks/useGridContextMenu.ts: 274 lines
  - Context menu + bulk dialog state
  - getContextMenuPosition, handleRightClick, handleVoiceSelect
  - handleVoiceLabelClick, handleBulkPatternSelect
  - Click-outside + keyboard shortcut effects

useDrumGrid.ts coordinates the two sub-hooks and adds the isDragging
guard to handleLeftClick, keeping React Rules of Hooks intact.

All 239 tests pass, TypeScript clean.
