+++ src/hooks/useDrumGrid.ts
@@ -2,13 +2,15 @@
  * useDrumGrid Hook
  *
  * Shared logic for DrumGrid and DrumGridDark components.
- * Handles voice selection, drag painting, touch support, context menus, and bulk operations.
+ * Coordinates voice selection state, drag painting, and context menu sub-hooks.
  */
 
-import { useState, useRef, useEffect, useCallback } from 'react';
+import { useState, useCallback } from 'react';
 import { GrooveData, DrumVoice } from '../types';
 import { GrooveUtils, HI_HAT_PATTERNS, SNARE_PATTERNS, KICK_PATTERNS, BulkPattern } from '../core';
 import { DRUM_ROWS, DrumRow } from '../core/DrumVoiceConfig';
+import { useGridDragPaint } from './useGridDragPaint';
+import { useGridContextMenu } from './useGridContextMenu';
 
 /** A single note change for batch operations */
 export interface NoteChange {
@@ -88,14 +90,13 @@ export function useDrumGrid({
   onPreview,
   advancedEditMode = false,
 }: UseDrumGridProps): UseDrumGridReturn {
-  // Calculate notes per measure for the global time signature
   const notesPerMeasure = GrooveUtils.calcNotesPerMeasure(
     groove.division,
     groove.timeSignature.beats,
     groove.timeSignature.noteValue
   );
 
-  // Voice selection state - tracks which voices are selected for each row at each position
+  // Voice selection state — shared between drag and context menu sub-hooks
   const [voiceSelections, setVoiceSelections] = useState<Record<string, DrumVoice[]>>(() => {
     const initial: Record<string, DrumVoice[]> = {};
     groove.measures.forEach((_, measureIndex) => {
@@ -108,54 +109,7 @@ export function useDrumGrid({
     return initial;
   });
 
-  // Context menu state
-  const [contextMenu, setContextMenu] = useState<ContextMenuState | null>(null);
-  const contextMenuRef = useRef<HTMLDivElement>(null);
-
-  // Drag state
-  const [isDragging, setIsDragging] = useState(false);
-  const [dragMode, setDragMode] = useState<'paint' | 'erase' | null>(null);
-  const [dragMeasureIndex, setDragMeasureIndex] = useState<number>(0);
-  const [dragSourceVoices, setDragSourceVoices] = useState<DrumVoice[] | null>(null);
-
-  // Touch state
-  const [touchStartTime, setTouchStartTime] = useState<number>(0);
-  const [touchMoved, setTouchMoved] = useState(false);
-
-  // Bulk operations dialog state
-  const [bulkDialog, setBulkDialog] = useState<BulkDialogState | null>(null);
-
-  const getContextMenuPosition = useCallback((
-    eventTarget: EventTarget,
-    rowIndex: number
-  ): Pick<ContextMenuState, 'x' | 'y' | 'placement'> => {
-    const target = eventTarget as HTMLElement;
-    const rect = target.getBoundingClientRect();
-    const menuWidth = 200;
-    const menuHeight = 34 + DRUM_ROWS[rowIndex].variations.length * 38 + 8;
-    const gap = 6;
-    const viewportPadding = 8;
-    const roomBelow = window.innerHeight - rect.bottom - gap - viewportPadding;
-    const roomAbove = rect.top - gap - viewportPadding;
-    const hasRoomBelow = roomBelow >= menuHeight;
-    const hasRoomAbove = roomAbove >= menuHeight;
-    const placement = hasRoomBelow || (!hasRoomAbove && roomBelow >= roomAbove) ? 'below' : 'above';
-
-    return {
-      x: Math.min(
-        Math.max(rect.left, viewportPadding),
-        window.innerWidth - menuWidth - viewportPadding
-      ),
-      y: placement === 'below'
-        ? rect.bottom + gap
-        : rect.top - menuHeight - gap,
-      placement,
-    };
-  }, []);
-
-  // ==========================================================================
-  // Helper functions
-  // ==========================================================================
+  // ===== Helper functions =====
 
   const isDownbeat = useCallback((pos: number) => {
     const notesPerBeat = groove.division / groove.timeSignature.beats;
@@ -213,375 +167,48 @@ export function useDrumGrid({
     return [];
   }, []);
 
-  // ==========================================================================
-  // Drag operations
-  // ==========================================================================
+  // ===== Sub-hooks =====
 
-  const applyDragAction = useCallback((
-    mode: 'paint' | 'erase',
-    measureIndex: number,
-    rowIndex: number,
-    position: number,
-    sourceVoices?: DrumVoice[]
-  ) => {
-    const measure = groove.measures[measureIndex];
-    if (!measure) return;
+  const ctx = useGridContextMenu({
+    groove,
+    onNoteToggle,
+    onSetNotes,
+    onPreview,
+    advancedEditMode,
+    getVoicesForPosition,
+    isPositionActive,
+    setVoiceSelections,
+  });
 
-    const isActive = isPositionActive(measureIndex, rowIndex, position);
+  const drag = useGridDragPaint({
+    groove,
+    onNoteToggle,
+    getVoicesForPosition,
+    isPositionActive,
+    setVoiceSelections,
+    setContextMenu: ctx.setContextMenu,
+    getContextMenuPosition: ctx.getContextMenuPosition,
+  });
 
-    if (mode === 'paint' && !isActive) {
-      const voices = sourceVoices || getVoicesForPosition(measureIndex, rowIndex, position);
-      voices.forEach(voice => {
-        onNoteToggle(voice, position, measureIndex);
-      });
-      if (sourceVoices) {
-        const key = `${measureIndex}-${rowIndex}-${position}`;
-        setVoiceSelections(prev => ({ ...prev, [key]: sourceVoices }));
-      }
-    } else if (mode === 'erase' && isActive) {
-      const row = DRUM_ROWS[rowIndex];
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[position]) {
-            onNoteToggle(voice, position, measureIndex);
-          }
-        });
-      });
-    }
-  }, [groove.measures, isPositionActive, getVoicesForPosition, onNoteToggle]);
-
-  // ==========================================================================
-  // Event handlers
-  // ==========================================================================
-
-  const handleMouseDown = useCallback((
-    event: React.MouseEvent,
-    measureIndex: number,
-    rowIndex: number,
-    position: number
-  ) => {
-    if (event.ctrlKey || event.metaKey) {
-      event.preventDefault();
-      const sourceVoices = getVoicesForPosition(measureIndex, rowIndex, position);
-      setIsDragging(true);
-      setDragMode('paint');
-      setDragMeasureIndex(measureIndex);
-      setDragSourceVoices(sourceVoices);
-      applyDragAction('paint', measureIndex, rowIndex, position, sourceVoices);
-    } else if (event.altKey || event.shiftKey) {
-      event.preventDefault();
-      setIsDragging(true);
-      setDragMode('erase');
-      setDragMeasureIndex(measureIndex);
-      setDragSourceVoices(null);
-      applyDragAction('erase', measureIndex, rowIndex, position);
-    }
-  }, [getVoicesForPosition, applyDragAction]);
-
-  const handleMouseEnter = useCallback((
-    measureIndex: number,
-    rowIndex: number,
-    position: number
-  ) => {
-    if (isDragging && dragMode && measureIndex === dragMeasureIndex) {
-      applyDragAction(dragMode, measureIndex, rowIndex, position, dragSourceVoices || undefined);
-    }
-  }, [isDragging, dragMode, dragMeasureIndex, dragSourceVoices, applyDragAction]);
-
-  const handleTouchStart = useCallback((
-    _event: React.TouchEvent,
-    measureIndex: number,
-    rowIndex: number,
-    position: number
-  ) => {
-    setTouchStartTime(Date.now());
-    setTouchMoved(false);
-    const sourceVoices = getVoicesForPosition(measureIndex, rowIndex, position);
-    setIsDragging(true);
-    setDragMode('paint');
-    setDragMeasureIndex(measureIndex);
-    setDragSourceVoices(sourceVoices);
-    applyDragAction('paint', measureIndex, rowIndex, position, sourceVoices);
-  }, [getVoicesForPosition, applyDragAction]);
-
-  const handleTouchMove = useCallback((event: React.TouchEvent) => {
-    setTouchMoved(true);
-    if (!isDragging) return;
-
-    const touch = event.touches[0];
-    const element = document.elementFromPoint(touch.clientX, touch.clientY);
-
-    // Check for both class names used by the two grid components
-    if (element && (element.classList.contains('note-cell') || element.classList.contains('drum-cell'))) {
-      const cellButton = element as HTMLButtonElement;
-      const mIdx = parseInt(cellButton.dataset.measureIndex || '-1');
-      const rIdx = parseInt(cellButton.dataset.rowIndex || '-1');
-      const pos = parseInt(cellButton.dataset.position || '-1');
-
-      if (mIdx >= 0 && rIdx >= 0 && pos >= 0 && dragMode && mIdx === dragMeasureIndex) {
-        applyDragAction(dragMode, mIdx, rIdx, pos, dragSourceVoices || undefined);
-      }
-    }
-  }, [isDragging, dragMode, dragMeasureIndex, dragSourceVoices, applyDragAction]);
-
-  const handleTouchEnd = useCallback((
-    event: React.TouchEvent,
-    measureIndex: number,
-    rowIndex: number,
-    position: number
-  ) => {
-    const touchDuration = Date.now() - touchStartTime;
-
-    if (touchDuration > 500 && !touchMoved) {
-      event.preventDefault();
-      const row = DRUM_ROWS[rowIndex];
-      if (row.variations.length > 1) {
-        // Get touch position for context menu
-        const menuPosition = getContextMenuPosition(event.currentTarget, rowIndex);
-        setContextMenu({
-          visible: true,
-          ...menuPosition,
-          rowIndex,
-          position,
-          measureIndex,
-        });
-      }
-    }
-
-    setIsDragging(false);
-    setDragMode(null);
-    setDragSourceVoices(null);
-    setTouchMoved(false);
-  }, [touchStartTime, touchMoved, getContextMenuPosition]);
-
+  // Guard left-click against ongoing drag (ctx.handleLeftClick doesn't own isDragging state)
   const handleLeftClick = useCallback((
     event: React.MouseEvent,
     measureIndex: number,
     rowIndex: number,
     position: number
   ) => {
-    if (isDragging) return;
-
-    const measure = groove.measures[measureIndex];
-    if (!measure) return;
-
-    if (advancedEditMode) {
-      // In advanced mode, left-click shows context menu
-      event.preventDefault();
-      const row = DRUM_ROWS[rowIndex];
-      if (row.variations.length > 1) {
-        const menuPosition = getContextMenuPosition(event.currentTarget, rowIndex);
-        setContextMenu({
-          visible: true,
-          ...menuPosition,
-          rowIndex,
-          position,
-          measureIndex,
-        });
-      }
-      return;
-    }
-
-    const voices = getVoicesForPosition(measureIndex, rowIndex, position);
-    const isActive = isPositionActive(measureIndex, rowIndex, position);
-
-    if (isActive) {
-      // Turn off all voices for this row at this position
-      const row = DRUM_ROWS[rowIndex];
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[position]) {
-            onNoteToggle(voice, position, measureIndex);
-          }
-        });
-      });
-    } else {
-      // First, clear any existing notes for this row at this position
-      const row = DRUM_ROWS[rowIndex];
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[position]) {
-            onNoteToggle(voice, position, measureIndex);
-          }
-        });
-      });
-      // Then turn on selected voices
-      voices.forEach(voice => {
-        onNoteToggle(voice, position, measureIndex);
-      });
-      onPreview(voices[0]);
-    }
-  }, [isDragging, groove.measures, advancedEditMode, getContextMenuPosition, getVoicesForPosition, isPositionActive, onNoteToggle, onPreview]);
-
-  const handleRightClick = useCallback((
-    event: React.MouseEvent,
-    measureIndex: number,
-    rowIndex: number,
-    position: number
-  ) => {
-    event.preventDefault();
-    const row = DRUM_ROWS[rowIndex];
-    if (row.variations.length > 1) {
-      const menuPosition = getContextMenuPosition(event.currentTarget, rowIndex);
-      setContextMenu({
-        visible: true,
-        ...menuPosition,
-        rowIndex,
-        position,
-        measureIndex,
-      });
-    }
-  }, [getContextMenuPosition]);
-
-  const handleVoiceSelect = useCallback((voices: DrumVoice[]) => {
-    if (!contextMenu) return;
-
-    const { measureIndex, rowIndex, position } = contextMenu;
-    const measure = groove.measures[measureIndex];
-    if (!measure) return;
-
-    const key = `${measureIndex}-${rowIndex}-${position}`;
-    setVoiceSelections(prev => ({ ...prev, [key]: voices }));
-
-    if (onSetNotes) {
-      const changes: NoteChange[] = [];
-      const row = DRUM_ROWS[rowIndex];
-
-      // Clear any existing notes for this row at this position
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[position]) {
-            changes.push({ voice, position, measureIndex, value: false });
-          }
-        });
-      });
-
-      // Set the new voices
-      voices.forEach(voice => {
-        changes.push({ voice, position, measureIndex, value: true });
-      });
-
-      onSetNotes(changes);
-    } else {
-      // Fallback to individual toggles
-      const row = DRUM_ROWS[rowIndex];
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[position]) {
-            onNoteToggle(voice, position, measureIndex);
-          }
-        });
-      });
-      voices.forEach(voice => {
-        onNoteToggle(voice, position, measureIndex);
-      });
-    }
-
-    onPreview(voices[0]);
-    setContextMenu(null);
-  }, [contextMenu, groove.measures, onSetNotes, onNoteToggle, onPreview]);
-
-  const handleVoiceLabelClick = useCallback((rowIndex: number, measureIndex: number) => {
-    setBulkDialog({ visible: true, rowIndex, measureIndex });
-  }, []);
-
-  const handleBulkPatternSelect = useCallback((pattern: BulkPattern) => {
-    if (!bulkDialog) return;
-
-    const { rowIndex, measureIndex } = bulkDialog;
-    const measure = groove.measures[measureIndex];
-    if (!measure) return;
-
-    const row = DRUM_ROWS[rowIndex];
-
-    // First, clear all voices for this row in this measure
-    for (let pos = 0; pos < notesPerMeasure; pos++) {
-      row.variations.forEach(v => {
-        v.voices.forEach(voice => {
-          if (measure.notes[voice]?.[pos]) {
-            onNoteToggle(voice, pos, measureIndex);
-          }
-        });
-      });
-    }
-
-    // Then, apply the pattern
-    for (let pos = 0; pos < notesPerMeasure; pos++) {
-      const shouldBeOn = pattern.pattern(pos, notesPerMeasure);
-      if (shouldBeOn && pattern.voices.length > 0) {
-        pattern.voices.forEach(voice => {
-          onNoteToggle(voice, pos, measureIndex);
-        });
-      }
-    }
-
-    if (pattern.voices.length > 0) {
-      onPreview(pattern.voices[0]);
-    }
-  }, [bulkDialog, groove.measures, notesPerMeasure, onNoteToggle, onPreview]);
-
-  // ==========================================================================
-  // Effects
-  // ==========================================================================
-
-  // Close context menu when clicking outside or handle keyboard shortcuts
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (contextMenuRef.current && !contextMenuRef.current.contains(event.target as Node)) {
-        setContextMenu(null);
-      }
-    };
-
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (!contextMenu?.visible) return;
-
-      const row = DRUM_ROWS[contextMenu.rowIndex];
-      const variation = row.variations.find(v => v.shortcut === event.key);
-
-      if (variation) {
-        event.preventDefault();
-        handleVoiceSelect(variation.voices);
-      } else if (event.key === 'Escape') {
-        setContextMenu(null);
-      }
-    };
-
-    if (contextMenu?.visible) {
-      document.addEventListener('mousedown', handleClickOutside);
-      document.addEventListener('keydown', handleKeyDown);
-      return () => {
-        document.removeEventListener('mousedown', handleClickOutside);
-        document.removeEventListener('keydown', handleKeyDown);
-      };
-    }
-  }, [contextMenu, handleVoiceSelect]);
-
-  // Handle global mouseup to end drag operations
-  useEffect(() => {
-    const handleMouseUp = () => {
-      if (isDragging) {
-        setIsDragging(false);
-        setDragMode(null);
-        setDragSourceVoices(null);
-      }
-    };
-
-    document.addEventListener('mouseup', handleMouseUp);
-    return () => {
-      document.removeEventListener('mouseup', handleMouseUp);
-    };
-  }, [isDragging]);
+    if (drag.isDragging) return;
+    ctx.handleLeftClick(event, measureIndex, rowIndex, position);
+  }, [drag.isDragging, ctx.handleLeftClick]);
 
   return {
-    // State
     voiceSelections,
-    contextMenu,
-    isDragging,
-    dragMode,
-    bulkDialog,
-    contextMenuRef: contextMenuRef as React.RefObject<HTMLDivElement>,
+    contextMenu: ctx.contextMenu,
+    isDragging: drag.isDragging,
+    dragMode: drag.dragMode,
+    bulkDialog: ctx.bulkDialog,
+    contextMenuRef: ctx.contextMenuRef,
 
-    // Helper functions
     isDownbeat,
     getVoicesForPosition,
     isPositionActive,
@@ -590,21 +217,19 @@ export function useDrumGrid({
     getPositionsForMeasure,
     getBulkPatternsForRow,
 
-    // Event handlers
-    handleMouseDown,
-    handleMouseEnter,
-    handleTouchStart,
-    handleTouchMove,
-    handleTouchEnd,
+    handleMouseDown: drag.handleMouseDown,
+    handleMouseEnter: drag.handleMouseEnter,
+    handleTouchStart: drag.handleTouchStart,
+    handleTouchMove: drag.handleTouchMove,
+    handleTouchEnd: drag.handleTouchEnd,
     handleLeftClick,
-    handleRightClick,
-    handleVoiceSelect,
-    handleVoiceLabelClick,
-    handleBulkPatternSelect,
+    handleRightClick: ctx.handleRightClick,
+    handleVoiceSelect: ctx.handleVoiceSelect,
+    handleVoiceLabelClick: ctx.handleVoiceLabelClick,
+    handleBulkPatternSelect: ctx.handleBulkPatternSelect,
 
-    // State setters
-    setContextMenu,
-    setBulkDialog,
+    setContextMenu: ctx.setContextMenu,
+    setBulkDialog: ctx.setBulkDialog,
   };
 }
